Seperate 'cache_hybrid' file from 'cache' file
