+++ storage/statedb/cache.go
@@ -52,42 +52,3 @@ func NewCache(cacheType CacheType, maxBytes int, redisEndpoint []string, redisCl
 	}
 	return nil, errNotSupportedCacheType
 }
-
-func NewHybridCache(maxBytes int, endpoint []string, isCluster bool) (Cache, error) {
-	redis, err := NewRedisCache(endpoint, isCluster)
-	if err != nil {
-		return nil, err
-	}
-	return &hybridCache{
-		local:  NewFastCache(maxBytes),
-		remote: redis,
-	}, nil
-}
-
-// hybridCache integrates two kinds of caches: local, remote.
-// local cache uses memory of the local machine and remote cache uses memory of the remote machine.
-type hybridCache struct {
-	local  Cache
-	remote Cache
-}
-
-func (cache *hybridCache) Set(k, v []byte) {
-	cache.local.Set(k, v)
-	cache.remote.Set(k, v)
-}
-
-func (cache *hybridCache) Get(k []byte) []byte {
-	ret := cache.local.Get(k)
-	if ret != nil {
-		return ret
-	}
-	return cache.remote.Get(k)
-}
-
-func (cache *hybridCache) Has(k []byte) ([]byte, bool) {
-	ret, has := cache.local.Has(k)
-	if has {
-		return ret, has
-	}
-	return cache.remote.Has(k)
-}
